+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - (CSS) 🚀 New MacOS-like Cupertino theme in light and dark variants.
 - (CSS) 🚀 New [Dracula](https://ui.draculatheme.com/) theme.
+- (CSS) 🚀 New `TabPane` style. There are three styles supported: default, floating and classic (new one).
 
 ### Improvements